refactor(core): dedup doc_sparse + fit_gamma into variational (#74 A2)
Move the duplicated sparse-bag and pooled-ridge kernels into the shared
variational module:

- `variational/sparse.rs::doc_sparse` — the BTreeMap form (deterministic
  sorted word order); ctm.rs and sts.rs drop their private copies, etm.rs
  re-points its import, python.rs calls the canonical path.
- `variational/mstep.rs::fit_gamma_ridge` — the flat-xtl pooled ridge solve
  `(X'X + ridge·I)⁻¹ X'Λ`; unifies ctm's `fit_gamma` and sts's
  `fit_gamma_ridge` (verified numerically identical). ctm's `fit_gamma_enet`
  (CTM-only lasso path) is left in place.

Pure extraction, no math changed. Gate: cargo test --lib 60 passed,
pytest 1917 passed/18 skipped, mkdocs --strict clean.
