+++ src/etm.rs
@@ -28,8 +28,8 @@
 //! SGD) for corpora too large for the per-document E-step; it would reproduce the
 //! reference's scaling at some cost in per-document posterior accuracy.
 
-use crate::ctm::{ctm_grad, ctm_hpb, ctm_lhood, doc_sparse, HpbResult};
-use crate::variational::lbfgs_minimize;
+use crate::ctm::{ctm_grad, ctm_hpb, ctm_lhood, HpbResult};
+use crate::variational::{lbfgs_minimize, doc_sparse};
 use crate::linalg::{cholesky, half_logdet, make_diagonally_dominant, spd_inverse};
 use rand::Rng;
 use rayon::prelude::*;
